add filteredChallenge

+++ src/services/challengeService.ts
@@ -4,7 +4,8 @@ import { UpdateChallengeDTO } from "../DTO/Challenge/UpdateChallengeDTO";
 import { Challenge } from "../entities/Challenge";
 import { TypeExercice } from "../entities/TypeExercice";
 import { User } from "../entities/User";
-import { In } from "typeorm";
+import { Equal, In } from "typeorm";
+import { DifficultyLevel } from "../enums/DifficultyLevel";
 
 const challengeRepository = AppDataSource.getRepository(Challenge);
 const userRepository = AppDataSource.getRepository(User);
@@ -42,11 +43,15 @@ export const createChallenge = async (dto: CreateChallengeDTO) => {
     return challengeRepository.save(challenge);
 }
 
-export const updateChallenge = async (challengeId: number, dto: UpdateChallengeDTO) => {
+export const updateChallenge = async (challengeId: number, dto: UpdateChallengeDTO, currentUserId: string) => {
     const challenge = await challengeRepository.findOne({ 
-         where: {id: challengeId }, relations: ["recommendedExercises", "participants"]})    
+         where: {id: challengeId }, relations: ["recommendedExercises", "participants", "creator"]})    
     if (!challenge) {
         throw new Error("Aucun challenge trouvé");
+    }
+
+    if(challenge.creator.id !== currentUserId) {
+        throw new Error("Vous ne pouvez pas modifié ce challenge");
     }
 
     let exercices: TypeExercice[] = [];
@@ -65,6 +70,14 @@ export const updateChallenge = async (challengeId: number, dto: UpdateChallengeD
     if (dto.durationMinutes !== undefined) {challenge.durationMinutes = dto.durationMinutes }
     if (dto.recommendedExercises !== undefined) {challenge.recommendedExercises = exercices }
     if (dto.participants !== undefined) {challenge.participants = participants }
+    if(dto.difficulty !== undefined) {
+        if(!Object.values(DifficultyLevel).includes(dto.difficulty)) {
+            throw new Error("Difficulté invalide");
+        }
+        challenge.difficulty = dto.difficulty
+    }
+    
+    
     const res = await challengeRepository.save(challenge);
     return res;
 }
@@ -77,8 +90,17 @@ export const findChallengeId = async (challengeId: number) => {
     return challengeRepository.findOne({ where: { id: challengeId }, relations: ["creator", "recommendedExercises", "participants"] });
 }
 
-export const deleteChallenge = async (challengeId: number) => {
+export const deleteChallenge = async (challengeId: number, currentUserId: string) => {
 
+    const challenge = await challengeRepository.findOne({ 
+         where: {id: challengeId }, relations: ["creator"]})    
+    if (!challenge) {
+        throw new Error("Aucun challenge trouvé");
+    }
+
+    if(challenge.creator.id !== currentUserId) {
+        throw new Error("Vous ne pouvez pas supprimé ce challenge ");
+    }
     return challengeRepository.delete(challengeId);
 }
 
@@ -104,7 +126,26 @@ export const shareChallenge = async (challengeId: number, participantsIds: numbe
             })
         }
     return challengeRepository.save(challenge);
-
 }
 
+
+export const filterChallenges = async(difficulty?: DifficultyLevel, duration?: number, typeExerciceNames?: string[]) => {
+    const query: any = {
+        where: {},
+        relations:["typesExercice"],
+    }
+
+    if(difficulty) {
+        query.where["difficulty"] = difficulty
+    }
+    if(duration !== undefined) {
+        query.where["durationMinutes"] = Equal(duration)
+    }
+
+    if(typeExerciceNames && typeExerciceNames.length > 0) {
+        query.where["typesExercice"] = { name: In(typeExerciceNames)}
+    }
+    const challenges = await challengeRepository.find(query)
+    return challenges
+}
